last polishing touches

+++ resources/js/components/AllBlogsComponent.vue
@@ -2,6 +2,8 @@
 
 import $ from 'jquery';
 import datatable from 'datatables.net';
+import moment from "moment";
+import Swal from "sweetalert2";
 
 export default {
     data() {
@@ -20,23 +22,35 @@ export default {
         });
 
         $(document).on('click','.delete_post', function () {
+
             let data_id = $(this).data('id');
-            // console.log(data_id);
             let data = {'post_id': data_id};
-            // console.log('Overview', data_id);
-            axios.post('/blog-data-delete', data)
-         .then(response => {
-             // var rowToDelete = $('#blog-table').DataTable().row(`#row-${id}`);
-             // rowToDelete.remove();
-             // $('#blog-table').DataTable().draw();
-             // this.blog
-             //ovde treba filter
-             self.deleteTable();
-             self.initTable();
-         })
-         .catch(error => {
-             console.log(error);
-         });
+
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    axios.post('/blog-data-delete', data)
+                        .then(response => {
+                            self.deleteTable();
+                            self.initTable();
+                        })
+                        .catch(error => {
+                            console.log(error);
+                        });
+                    Swal.fire(
+                        'Deleted!',
+                        'Your file has been deleted.',
+                        'success'
+                    )
+                }
+            })
 
 
         });
@@ -70,12 +84,39 @@ export default {
                     { title: 'ID', data: 'id' },
                     { title: 'Author', data: 'author' },
                     { title: 'Image', data: 'image', "render": function (data, type, row, meta) {
-                            return '<a href="storage/uploads/' + row.id + '/' + data + '"><img src="storage/uploads/' + row.id + '/' + data + '" class="img-thumbnail"></a>';
+                            return '<a href="storage/uploads/' + row.id + '/' + data + '"><img src="storage/uploads/' + row.id + '/' + data + '" class="img-thumbnail slika"></a>';
                         },
                     },
-                    { title: 'Created At', data: 'created_at' },
-                    { title: 'Published At', data: 'published_at' },
-                    { title: 'Archived At', data: 'archived_at' },
+                    { title: 'Created At',
+                        data: 'created_at',
+                        render: function(data, type, row) {
+                        if(data != null) {
+                            return '<span class="created_at" title="' + data + '">' + moment(data).format('d.m.Y') + '</span>';
+                        }
+                        else
+                            return data;
+                        },
+                    },
+                    { title: 'Published At',
+                        data: 'published_at',
+                        render: function(data, type, row) {
+                        if(data != null) {
+                            return '<span class="published_at" title="' + data + '">' + moment(data).format('d.m.Y') + '</span>';
+                        }
+                        else
+                            return data;
+                        },
+                    },
+                    { title: 'Archived At',
+                        data: 'archived_at',
+                        render: function(data, type, row) {
+                            if (data != null) {
+                                return '<span class="archived_at" title="' + data + '">' + moment(data).format('d.m.Y') + '</span>';
+                             }
+                            else
+                                return data;
+                        },
+                    },
                     { title: 'Title', data: 'title' },
                     { title: 'Status', data: 'status' },
                     { title: 'Type', data: 'type' },
@@ -86,7 +127,7 @@ export default {
                         render: function (data, type, row) {
                             return `  <div class="d-flex justify-content-around">
                                 <a href="/blog-data-update/${data.id}"class="btn btn-sm btn-primary edit_post" data-id="${data.id}">Edit</a>
-                                <button class="btn btn-sm btn-danger ml-3 pl-3 delete_post" data-id="${data.id}">Delete</button>
+                                <button class="btn btn-sm btn-danger ml-3 pl-3 delete_post " data-id="${data.id}">Delete</button>
                             </div>
                             `;
                         },
@@ -94,10 +135,6 @@ export default {
 
                 ],
             });
-
-
-
-
         },
         deleteTable()
         {
@@ -142,5 +179,11 @@ table.dataTable tbody tr {
 
 table.dataTable.stripe tbody tr.odd, table.dataTable.display tbody tr.odd {
     background-color: #999292 !important;
+}
+
+.slika
+{
+    height: 40px;
+    width: 80px;
 }
 </style>
